+++ lib/acmex/client.ex
@@ -1,14 +1,25 @@
 defmodule Acmex.Client do
-  @moduledoc false
+  @moduledoc """
+  This module is responsible for calling the ACME API.
+
+  `GenServer` is used to store the account and its credentials.
+  """
 
   use GenServer
 
   alias Acmex.{Crypto, Request}
   alias Acmex.Resource.{Account, Authorization, Challenge, Directory, Order}
 
+  @bad_nonce_error_type "urn:ietf:params:acme:error:badNonce"
+
+  @doc """
+  Starts the `GenServer` with a given account key.
+  """
+  @spec start_link(String.t(), GenServer.name()) :: GenServer.on_start()
   def start_link(key, name \\ __MODULE__),
     do: GenServer.start_link(__MODULE__, [key: key], name: name)
 
+  @impl true
   def init(key: key) do
     with {:ok, jwk} <- Crypto.fetch_jwk_from_key(key),
          {:ok, directory} <- Directory.new() do
@@ -24,6 +35,7 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call({:new_account, contact, terms_of_service_agreed}, _from, state) do
     payload = %{contact: contact, termsOfServiceAgreed: terms_of_service_agreed}
 
@@ -33,6 +45,7 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call(:get_account, _from, state) do
     case get_account(state.account, state.directory, state.jwk) do
       {:ok, account} -> {:reply, {:ok, account}, %{state | account: account}}
@@ -40,6 +53,7 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call({:new_order, identifiers}, _from, state) do
     payload = %{identifiers: Enum.map(identifiers, &Map.new(type: "dns", value: &1))}
 
@@ -52,6 +66,7 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call({:get_order, url}, _from, state) do
     with {:ok, resp} <- post_as_get(url, state),
          order <- Order.new(resp.body, resp.headers),
@@ -63,6 +78,7 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call({:get_challenge, url}, _from, state) do
     with {:ok, %{body: body}} <- post_as_get(url, state),
          challenge <- Challenge.new(body),
@@ -73,9 +89,11 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call({:get_challenge_response, challenge}, _from, %{jwk: jwk} = state),
     do: {:reply, Challenge.get_response(challenge, jwk), state}
 
+  @impl true
   def handle_call({:validate_challenge, challenge}, _from, state) do
     {:ok, key_authorization} = Challenge.get_key_authorization(challenge, state.jwk)
     payload = %{key_authorization: key_authorization}
@@ -86,6 +104,7 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call({:finalize_order, order, csr}, _from, state) do
     payload = %{csr: Base.url_encode64(csr, padding: false)}
 
@@ -95,6 +114,7 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call({:get_certificate, order}, _from, state) do
     case post_as_get(order.certificate, state, [{"Accept", "application/pem-certificate-chain"}]) do
       {:ok, %{body: body}} -> {:reply, {:ok, body}, state}
@@ -102,6 +122,7 @@ defmodule Acmex.Client do
     end
   end
 
+  @impl true
   def handle_call({:revoke_certificate, certificate, reason}, _from, state) do
     [{:Certificate, der_certificate, _enc}, _] = :public_key.pem_decode(certificate)
     payload = %{certificate: Base.url_encode64(der_certificate, padding: false), reason: reason}
@@ -123,7 +144,7 @@ defmodule Acmex.Client do
       url = Request.get_header(resp.headers, "Location")
       {:ok, Account.new(Map.put(resp.body, :url, url))}
     else
-      {:error, %HTTPoison.Response{body: %{type: "urn:ietf:params:acme:error:badNonce"}}} ->
+      {:error, %HTTPoison.Response{body: %{type: @bad_nonce_error_type}}} ->
         new_account(directory, jwk, payload)
 
       error ->
@@ -161,7 +182,7 @@ defmodule Acmex.Client do
          {:ok, resp} <- Request.post(url, state.jwk, payload, nonce, kid) do
       {:ok, resp}
     else
-      {:error, %HTTPoison.Response{body: %{type: "urn:ietf:params:acme:error:badNonce"}}} ->
+      {:error, %HTTPoison.Response{body: %{type: @bad_nonce_error_type}}} ->
         post(url, state, payload)
 
       error ->
@@ -174,7 +195,7 @@ defmodule Acmex.Client do
          {:ok, resp} <- Request.post_as_get(url, state.jwk, nonce, kid, headers) do
       {:ok, resp}
     else
-      {:error, %HTTPoison.Response{body: %{type: "urn:ietf:params:acme:error:badNonce"}}} ->
+      {:error, %HTTPoison.Response{body: %{type: @bad_nonce_error_type}}} ->
         post_as_get(url, state, headers)
 
       error ->