+++ test/acmex/resource/directory_test.exs
@@ -4,18 +4,13 @@ defmodule Acmex.Resource.DirectoryTest do
   alias Acmex.Config
   alias Acmex.Resource.Directory
 
-  describe "Directory.new/1" do
-    test "returns directory struct" do
-      {:ok, directory} = Directory.new(Config.directory_url())
-
-      assert directory.__struct__ == Directory
-      assert directory.new_account
+  describe "new/1" do
+    test "returns a directory struct" do
+      assert {:ok, %Directory{}} = Directory.new(Config.directory_url())
     end
 
-    test "returns error" do
-      {:error, reason} = Directory.new("#{Config.directory_url()}/invalid")
-
-      assert reason.status_code == 404
+    test "when directory URL is invalid, returns an error" do
+      assert {:error, %{status_code: 404}} = Directory.new("#{Config.directory_url()}/invalid")
     end
   end
 end