+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Improve the documentation.
+
+### Fixed
+- Avoid to start `Acmex` when the key is invalid. An error should be returned.
+
 ## [0.1.1] - 2020-04-26
 ### Changed
 - Update dependencies.